refactor: 개인 맞춤 전략 컴포넌트 -> 페이지
