+++ src/components/AssetPlanStrategy/StrategyForPrivate.vue
