+++ src/components/AssetPlanStrategy/StrategyForPrivate/StrategyCircle.vue
