+++ src/components/AssetPlanStrategy/StrategyForPrivate/StrategyText.vue
