+++ src/router/router.js
@@ -16,6 +16,7 @@ import WithdrawMember from "../components/MyPage/WithdrawMember.vue";
 import KakaoRedirectPage from "../pages/RedirectPages/KakaoRedirectPage.vue";
 import AssetAnalyze from "../pages/AssetAnalyze.vue";
 import AllProducts from "../pages/AllProducts.vue";
+import AssetPlanStrategy from "../pages/AssetPlanStrategy.vue";
 
 const router=createRouter({
     history:createWebHistory(), //라우팅 모드 
@@ -30,6 +31,7 @@ const router=createRouter({
         {path:'/asset-analyze',component:AssetAnalyze},
 
         {path:'/asset-plan',component:AssetPlan},
+        {path:'/asset-plan/strategy',component:AssetPlanStrategy},
 
         {path:'/mypage',name: 'MyPage',component:MyPage,props:true},
         {path:'/login',name: 'Login',component:Login},
